requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.6
Django==4.2.4
djangorestframework==3.14.0
flake8==6.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
pytz==2023.3
sqlparse==0.4.4
drf-spectacular==0.26.4
